LayoutTests/http/tests/paymentrequest/ApplePayModifier-disbursementPaymentRequest.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>
    Tests for providing `disbursementPaymentRequest` as part of
    `ApplePayModifier`.
</title>
<script src="/js-test-resources/ui-helper.js"></script>
<script src="/resources/payment-request.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
    <script>
        setup({ explicit_done: true, explicit_timeout: true });

        test((test) => {
            const method = validPaymentMethod();
            try {
                method.data.features = ["invalidFeatureThatsNotDisbursements"];
                new PaymentRequest([method], validPaymentDetails());
                assert_unreached("Should have thrown with an invalid feature");
            } catch {
                method.data.features = ["paymentRequestDisbursements"];
                new PaymentRequest([method], validPaymentDetails());
            }
        }, "Should have a feature for `disbursementPaymentRequest`.");

        const RESOLVED_PROMISE = Promise.resolve({});

        const lineItems = {
            total: {
                label: "Chime",
                amount: "15.00",
            },
            ifo: {
                label: "Instant Transfer Fee",
                amount: "0.15",
                disbursementLineItemType: "instantFundsOutFee",
            },
            disbursement: {
                label: "Total Amount",
                amount: "14.85",
                disbursementLineItemType: "disbursement",
            },
        };

        const validLineItems = [
            lineItems.total,
            lineItems.ifo,
            lineItems.disbursement,
        ];

        function modifiersWithData(data, lineItems) {
            data.additionalLineItems = lineItems;
            return [
                {
                    supportedMethods: "https://apple.com/apple-pay",
                    data: { ...data },
                },
            ];
        }

        function testValidDisbursementPaymentRequest(description, options) {
            const {
                initialDisbursementPaymentRequest,
                expectedDisbursementPaymentRequest,
            } = options;
            user_activation_test(async (test) => {
                const detailsInit = {
                    total: {
                        label: "Disbursement",
                        amount: {
                            value: "14.85",
                            currency: "USD",
                        },
                    },
                };
                if (initialDisbursementPaymentRequest !== undefined) {
                    detailsInit.modifiers = modifiersWithData(
                        {
                            disbursementPaymentRequest:
                                initialDisbursementPaymentRequest,
                        },
                        validLineItems
                    );
                }

                const request = new PaymentRequest(
                    [validPaymentMethod()],
                    detailsInit
                );
                request.addEventListener("merchantvalidation", (event) => {
                    event.complete(RESOLVED_PROMISE);
                });

                request.addEventListener("shippingaddresschange", (event) => {
                    event.updateWith({});
                    internals.mockPaymentCoordinator.changePaymentMethod({
                        type: "debit",
                    });
                });

                request.addEventListener("paymentmethodchange", (event) => {
                    event.updateWith(RESOLVED_PROMISE);
                    internals.mockPaymentCoordinator.acceptPayment();
                });

                const response = await request.show();

                const { disbursementPaymentRequest } =
                    internals.mockPaymentCoordinator;
                const assert = disbursementPaymentRequest
                    ? assert_object_equals
                    : assert_equals;
                assert(
                    disbursementPaymentRequest,
                    expectedDisbursementPaymentRequest,
                    "check that the `DisbursementPaymentRequest` matches"
                );
                await response.complete("success");
            }, description + " as part of the request.");
        }

        function testInvalidDisbursementPaymentRequest(
            description,
            { disbursementPaymentRequest, expectedErrorType }
        ) {
            user_activation_test(async (test) => {
                const detailsInit = {
                    total: {
                        label: "Disbursement",
                        amount: {
                            value: "14.85",
                            currency: "USD",
                        },
                    },
                };

                if (disbursementPaymentRequest) {
                    detailsInit.modifiers = modifiersWithData(
                        { disbursementPaymentRequest },
                        validLineItems
                    );
                }

                const request = new PaymentRequest(
                    [validPaymentMethod()],
                    detailsInit
                );

                try {
                    await request.show();
                } catch (e) {
                    assert_true(
                        e instanceof expectedErrorType,
                        `Expected ${expectedErrorType}`
                    );
                }
            }, description + " as part of the request.");
        }

        testValidDisbursementPaymentRequest("Should propagate all data", {
            initialDisbursementPaymentRequest: {},
            expectedDisbursementPaymentRequest: {},
        });

        testValidDisbursementPaymentRequest(
            "Should propagate recipient contact information",
            {
                initialDisbursementPaymentRequest: {
                    requiredRecipientContactFields: ["email", "name"],
                },
                expectedDisbursementPaymentRequest: {
                    requiredRecipientContactFields: ["email", "name"],
                },
            }
        );

        testInvalidDisbursementPaymentRequest(
            "Should error when incorrect `requiredRecipientContactFields` are provided",
            {
                disbursementPaymentRequest: {
                    requiredRecipientContactFields: ["invalid1", "invalid2"],
                },
                expectedErrorType: TypeError,
            }
        );

        // We don't need invalidation tests since a disbursement request is an empty dictionary.

        done();
    </script>
</body>
